Point out brittleness in environment unit test
This isn't an actual pull request, but there wasn't a way to open a github issue on this repo.

I just wanted to point out that this test will not work if I set my SHELL env var to something other than /bin/bash.
This probably falls over in windows too. If you care about it, there is probably a solution.

Processbuilder allows you to specify an env for a subprocess, so you could try fiddling around with that. 

OR 
you could inject an EnvironmentVariableProvider Interface, to which you'd implement a fake in the unit test. Then you'd have a tiny implementation class as follows:
public class SystemEnvironmentVariableProvider implements EnvironmentVariableProvider {
   public string getEnv(String envVarName) { return System.getEnv(envVarName);} 
}